Improve QuillCode quality hotspots
