+++ E2E/playwright/tests/composer.spec.ts
@@ -100,343 +100,3 @@ test('mock harness changes approval mode independently from model selection', as
   await expect(page.getByTestId('mode-picker-button')).toHaveAttribute('data-mode-tone', 'read-only');
   await expect(page.getByTestId('model-picker-button')).not.toContainText('Read-only');
 });
-
-test('mock harness routes slash commands to workspace actions', async ({ page }) => {
-  await page.goto(harnessURL());
-
-  await page.getByLabel('Message').fill('/terminal');
-  await page.getByRole('button', { name: 'Send' }).click();
-  await expect(page.getByTestId('terminal-pane')).toBeVisible();
-  await expect(page.getByText('Terminal opened.')).toBeVisible();
-
-  await page.getByLabel('Message').fill('/browser');
-  await page.getByRole('button', { name: 'Send' }).click();
-  await expect(page.getByTestId('browser-pane')).toBeVisible();
-  await expect(page.getByText('Browser opened.')).toBeVisible();
-
-  await page.getByLabel('Message').fill('/diff');
-  await page.getByRole('button', { name: 'Send' }).click();
-  await expect(page.getByTestId('tool-card-title').last()).toHaveText('host.git.diff');
-
-  await page.getByLabel('Message').fill('/git-status');
-  await page.getByRole('button', { name: 'Send' }).click();
-  await expect(page.getByTestId('tool-card-title').last()).toHaveText('host.git.status');
-
-  await page.getByLabel('Message').fill('/worktrees');
-  await page.getByRole('button', { name: 'Send' }).click();
-  await expect(page.getByTestId('tool-card-title').last()).toHaveText('host.git.worktree.list');
-  await expect(page.getByTestId('tool-card-output').last()).toContainText('/mock/quillcode-existing');
-
-  await page.getByLabel('Message').fill('/worktree create slash-worktree --branch slash/demo --base main');
-  await page.getByRole('button', { name: 'Send' }).click();
-  await expect(page.getByTestId('project-item').first()).toContainText('slash-worktree');
-  await expect(page.getByTestId('top-bar-title')).toHaveText('Worktree: slash/demo');
-  await expect(page.getByTestId('message').last()).toContainText('Opened worktree slash-worktree at /mock/slash-worktree.');
-
-  await page.getByLabel('Message').fill('/worktree open slash-worktree');
-  await page.getByRole('button', { name: 'Send' }).click();
-  await expect(page.getByTestId('top-bar-title')).toHaveText('Worktree: slash-worktree');
-  await expect(page.getByTestId('message').last()).toContainText('Opened worktree slash-worktree at /mock/slash-worktree.');
-
-  await page.getByLabel('Message').fill('/worktree remove slash-worktree --force');
-  await page.getByRole('button', { name: 'Send' }).click();
-  await expect(page.getByTestId('tool-card-title').last()).toHaveText('host.git.worktree.remove');
-  await expect(page.getByTestId('tool-card-input').last()).toContainText('"force": true');
-  await expect(page.getByTestId('message').last()).toContainText('Removed worktree slash-worktree.');
-
-  await page.getByLabel('Message').fill('/worktree prune --dry-run --verbose');
-  await page.getByRole('button', { name: 'Send' }).click();
-  await expect(page.getByTestId('tool-card-title').last()).toHaveText('host.git.worktree.prune');
-  await expect(page.getByTestId('tool-card-input').last()).toContainText('"dryRun": true');
-  await expect(page.getByTestId('message').last()).toContainText('No stale worktree records found.');
-
-  await page.getByLabel('Message').fill('/pr');
-  await page.getByRole('button', { name: 'Send' }).click();
-  await expect(page.getByLabel('Message')).toHaveValue('Create a pull request titled ');
-
-  await page.getByLabel('Message').fill('/pr fill');
-  await page.getByRole('button', { name: 'Send' }).click();
-  await expect(page.getByTestId('tool-card-title').last()).toHaveText('host.git.pr.create');
-  await expect(page.getByTestId('tool-card-input').last()).toContainText('"fill": true');
-  await expect(page.getByTestId('message').last()).toContainText('Opened a pull request for the current branch');
-
-  await page.getByLabel('Message').fill('/compact');
-  await page.getByRole('button', { name: 'Send' }).click();
-  await expect(page.getByTestId('top-bar-title')).toContainText('Compact:');
-  await expect(page.getByTestId('message').first()).toContainText('Context compacted from');
-});
-
-test('mock harness suggests slash commands in the composer', async ({ page }) => {
-  await page.goto(harnessURL());
-
-  const message = page.getByLabel('Message');
-  await message.fill('/');
-  await expect(page.getByTestId('slash-suggestions')).toBeVisible();
-  await expect(page.getByTestId('slash-suggestion')).toHaveCount(6);
-  await expect(page.getByTestId('slash-suggestion').first()).toContainText('/help');
-  await expect(page.locator('[data-testid="slash-suggestion"][data-selected="true"]')).toContainText('/help');
-
-  await page.keyboard.press('ArrowDown');
-  await expect(page.locator('[data-testid="slash-suggestion"][data-selected="true"]')).toContainText('/status');
-
-  await message.fill('/workt');
-  await expect(page.getByTestId('slash-suggestion').first()).toContainText('/worktrees');
-  await page.keyboard.press('Enter');
-  await expect(message).toHaveValue('/worktrees');
-  await expect(message).toBeFocused();
-
-  await page.keyboard.press('Enter');
-  await expect(page.getByTestId('slash-suggestions')).toHaveCount(0);
-  await expect(page.getByTestId('tool-card-title').last()).toHaveText('host.git.worktree.list');
-
-  await message.fill('/worktree c');
-  await expect(page.getByTestId('slash-suggestion').first()).toContainText('/worktree create path');
-  await page.keyboard.press('Tab');
-  await expect(message).toHaveValue('/worktree create ');
-
-  await message.fill('/project r');
-  await page.keyboard.press('ArrowDown');
-  await expect(page.locator('[data-testid="slash-suggestion"][data-selected="true"]')).toContainText('/project rename name');
-  await page.keyboard.press('Tab');
-  await expect(message).toHaveValue('/project rename ');
-
-  await message.fill('/fol');
-  await expect(page.getByTestId('slash-suggestion').first()).toContainText('/follow-up when');
-  await page.keyboard.press('Enter');
-  await expect(message).toHaveValue('/follow-up in ');
-
-  await message.fill('/workspace-c');
-  await expect(page.getByTestId('slash-suggestion').first()).toContainText('/workspace-check when');
-  await page.keyboard.press('Enter');
-  await expect(message).toHaveValue('/workspace-check in ');
-
-  await message.fill('/workt');
-  await page.getByTestId('slash-suggestion').first().click();
-  await expect(message).toHaveValue('/worktrees');
-  await expect(message).toBeFocused();
-});
-
-test('mock harness suggests workspace files for @ mentions in the composer', async ({ page }) => {
-  await page.goto(harnessURL());
-
-  const message = page.getByLabel('Message');
-
-  // A bare @ surfaces workspace files AND directories, shortest path first.
-  await message.fill('@');
-  await expect(page.getByTestId('file-mention-suggestions')).toBeVisible();
-  await expect(page.getByTestId('file-mention-suggestion')).toHaveCount(4);
-  // The `Sources` directory (shortest path at the top score tier) leads, badged as a directory.
-  const firstMention = page.getByTestId('file-mention-suggestion').first();
-  await expect(firstMention).toContainText('Sources');
-  await expect(firstMention).toHaveAttribute('data-kind', 'directory');
-  await expect(page.locator('[data-testid="file-mention-suggestion"][data-selected="true"]')).toHaveAttribute('data-kind', 'directory');
-
-  // ArrowDown moves the active mention to the next entry (README.md).
-  await page.keyboard.press('ArrowDown');
-  await expect(page.locator('[data-testid="file-mention-suggestion"][data-selected="true"]')).toContainText('README.md');
-
-  // An email-style token is not treated as a mention.
-  await message.fill('ping name@example.com');
-  await expect(page.getByTestId('file-mention-suggestions')).toHaveCount(0);
-
-  // A name prefix ranks the matching file first.
-  await message.fill('look at @Agent');
-  await expect(page.getByTestId('file-mention-suggestion')).toHaveCount(1);
-  await expect(page.getByTestId('file-mention-suggestion').first()).toContainText('Agent.swift');
-
-  // Enter accepts the active mention, replacing the @token and adding a trailing space.
-  await page.keyboard.press('Enter');
-  await expect(message).toHaveValue('look at @Sources/Agent.swift ');
-  await expect(message).toBeFocused();
-  await expect(page.getByTestId('file-mention-suggestions')).toHaveCount(0);
-
-  // Tab completes the selected file mention.
-  await message.fill('read @READ');
-  await page.keyboard.press('Tab');
-  await expect(message).toHaveValue('read @README.md ');
-
-  // Click-to-insert accepts a file mention.
-  await message.fill('open @Agent');
-  await page.getByTestId('file-mention-suggestion').first().click();
-  await expect(message).toHaveValue('open @Sources/Agent.swift ');
-  await expect(message).toBeFocused();
-
-  // A directory mention inserts a trailing slash so the agent reads the directory.
-  await message.fill('open @Sour');
-  const dirRow = page.locator('[data-testid="file-mention-suggestion"][data-kind="directory"]').first();
-  await expect(dirRow).toContainText('Sources');
-  await dirRow.click();
-  await expect(message).toHaveValue('open @Sources/ ');
-
-  // Slash commands take precedence over file mentions.
-  await message.fill('/help');
-  await expect(page.getByTestId('file-mention-suggestions')).toHaveCount(0);
-  await expect(page.getByTestId('slash-suggestions')).toBeVisible();
-});
-
-test('mock harness boosts and badges changed files in @ mentions after a git status', async ({ page }) => {
-  await page.goto(harnessURL());
-  const message = page.getByLabel('Message');
-
-  // Before any git status: nothing is flagged changed (the `Sources` directory leads the
-  // bare-@ list on the shortest-path tiebreak, but no entry carries a Changed badge).
-  await message.fill('@');
-  await expect(page.locator('[data-testid="file-mention-suggestion"][data-changed="true"]')).toHaveCount(0);
-
-  // Run a git status, which reports Sources/App.swift as changed.
-  await message.fill('/git-status');
-  await page.getByRole('button', { name: 'Send' }).click();
-  await expect(page.getByTestId('tool-card-title').last()).toHaveText('host.git.status');
-
-  // Now @ floats the changed file to the top with a Changed badge ...
-  await message.fill('@');
-  const first = page.getByTestId('file-mention-suggestion').first();
-  await expect(first).toContainText('Sources/App.swift');
-  await expect(first).toHaveAttribute('data-changed', 'true');
-  await expect(first.getByTestId('file-mention-changed-badge')).toBeVisible();
-  // ... while an unchanged file that also matches is not flagged (cross-surface scoring parity).
-  await expect(page.locator('[data-testid="file-mention-suggestion"][data-path="Sources/Agent.swift"]'))
-    .toHaveAttribute('data-changed', 'false');
-});
-
-test('mock harness preserves a separate composer draft per thread', async ({ page }) => {
-  await page.goto(harnessURL());
-  const message = page.getByLabel('Message');
-
-  // Create thread A by sending a prompt.
-  await message.fill('alpha topic');
-  await message.press('Enter');
-  await expect(message).toBeEnabled();
-  await expect(page.getByTestId('sidebar-item').filter({ hasText: 'alpha topic' })).toBeVisible();
-
-  // New chat, then create thread B by sending a prompt.
-  await page.getByTestId('new-chat-button').click();
-  await message.fill('beta topic');
-  await message.press('Enter');
-  await expect(message).toBeEnabled();
-  await expect(page.getByTestId('sidebar-item').filter({ hasText: 'beta topic' })).toBeVisible();
-
-  // Leave an unsent draft in thread B.
-  await message.fill('work in progress for beta');
-
-  // Switching to thread A stashes B's draft and shows A's own (empty) draft.
-  await page.getByTestId('sidebar-item').filter({ hasText: 'alpha topic' }).click();
-  await expect(message).toHaveValue('');
-
-  // Switching back to thread B restores its unsent draft verbatim.
-  await page.getByTestId('sidebar-item').filter({ hasText: 'beta topic' }).click();
-  await expect(message).toHaveValue('work in progress for beta');
-});
-
-test('mock harness recalls sent messages with Up and Down', async ({ page }) => {
-  await page.goto(harnessURL());
-  const message = page.getByLabel('Message');
-
-  await message.fill('first prompt');
-  await message.press('Enter');
-  await expect(message).toBeEnabled();
-  await expect(message).toHaveValue('');
-
-  await message.fill('second prompt');
-  await message.press('Enter');
-  await expect(message).toBeEnabled();
-  await expect(message).toHaveValue('');
-
-  // Up recalls the newest message first, then older, clamping at the oldest.
-  await message.press('ArrowUp');
-  await expect(message).toHaveValue('second prompt');
-  await message.press('ArrowUp');
-  await expect(message).toHaveValue('first prompt');
-  await message.press('ArrowUp');
-  await expect(message).toHaveValue('first prompt');
-
-  // Down walks back toward newer entries, then restores the empty in-progress draft.
-  await message.press('ArrowDown');
-  await expect(message).toHaveValue('second prompt');
-  await message.press('ArrowDown');
-  await expect(message).toHaveValue('');
-
-  // Up does not hijack editing once the composer already has text.
-  await message.fill('half typed');
-  await message.press('ArrowUp');
-  await expect(message).toHaveValue('half typed');
-});
-
-test('mock harness searches and selects models from the composer', async ({ page }) => {
-  await page.goto(harnessURL());
-
-  await page.getByTestId('model-picker-button').click();
-  await expect(page.getByTestId('model-browser')).toBeVisible();
-  await expect(page.getByTestId('model-result-count')).toHaveText('6 models available');
-  // Socrates 1.1 leads the Recommended list as the currently strongest model.
-  await expect(page.getByTestId('model-option').first()).toContainText('Socrates 1.1');
-  await expect(page.getByTestId('model-option-summary').first()).toContainText('Leading frontier model');
-  // The selected default (Nike) stays expanded by default.
-  await expect(page.getByTestId('model-detail-button').nth(1)).toHaveAttribute('aria-expanded', 'true');
-  await expect(page.getByTestId('model-capability')).toContainText('Nike 1.0 is the fast default');
-  await expect(page.getByTestId('model-metadata-row').filter({ hasText: 'trustedrouter/fast' })).toBeVisible();
-  await expect(page.getByTestId('model-metadata-row').filter({ hasText: 'Current, Default, Recommended' })).toBeVisible();
-  await expect(page.getByTestId('model-option')).toHaveCount(6);
-
-  await page.getByTestId('model-detail-button').nth(0).click();
-  await expect(page.getByTestId('model-detail-button').nth(0)).toHaveAttribute('aria-expanded', 'true');
-  await expect(page.getByTestId('model-capability')).toContainText('Socrates 1.1 is the leading model');
-  await expect(page.getByTestId('model-metadata-row').filter({ hasText: 'tr/socrates' })).toBeVisible();
-
-  await page.getByTestId('model-detail-button').nth(2).click();
-  await expect(page.getByTestId('model-detail-button').nth(2)).toHaveAttribute('aria-expanded', 'true');
-  await expect(page.getByTestId('model-capability')).toContainText('Synth is the balanced model');
-  await expect(page.getByTestId('model-metadata-row').filter({ hasText: 'tr/synth' })).toBeVisible();
-
-  await page.getByTestId('model-search').fill('socrates');
-  await expect(page.getByTestId('model-result-count')).toHaveText('1 model for "socrates"');
-  await expect(page.getByTestId('model-option')).toHaveCount(1);
-  await expect(page.getByTestId('model-option')).toContainText('Socrates 1.1');
-  await page.getByTestId('model-search').fill('');
-
-  await page.getByTestId('model-favorite-button').nth(1).click();
-  await expect(page.getByTestId('model-browser')).toBeVisible();
-  await expect(page.getByTestId('model-category').first()).toContainText('Favorites');
-  await expect(page.getByTestId('model-option')).toHaveCount(7);
-  await expect(page.getByTestId('model-favorite-button').first()).toHaveAttribute('aria-label', 'Remove favorite model');
-
-  await page.getByTestId('model-search').fill('favorite');
-  await expect(page.getByTestId('model-category')).toHaveCount(1);
-  await expect(page.getByTestId('model-category')).toContainText('Favorites');
-  await expect(page.getByTestId('model-option')).toHaveCount(1);
-
-  await page.getByTestId('model-search').fill('moon k2');
-  await expect(page.getByTestId('model-option')).toHaveCount(1);
-  await expect(page.getByTestId('model-option')).toContainText('moonshotai/Kimi K2.6');
-
-  await page.getByTestId('model-option').click();
-  await expect(page.getByTestId('model-picker-button')).toHaveText('moonshotai/Kimi K2.6');
-  await expect(page.getByTestId('mode-pill')).toHaveText('Auto');
-  await expect(page.getByTestId('model-browser')).toHaveCount(0);
-
-  await page.getByTestId('model-picker-button').click();
-  await page.getByTestId('model-search').fill('not-a-model');
-  await expect(page.getByTestId('model-result-count')).toHaveText('0 models for "not-a-model"');
-  await expect(page.getByTestId('model-empty')).toBeVisible();
-  await page.getByTestId('model-clear-search').first().click();
-  await expect(page.getByTestId('model-search')).toBeFocused();
-  await expect(page.getByTestId('model-result-count')).toHaveText('7 models available');
-});
-
-test('mock harness supports keyboard navigation in the model picker', async ({ page }) => {
-  await page.goto(harnessURL());
-
-  await page.getByTestId('model-picker-button').click();
-  await expect(page.getByTestId('model-search')).toBeFocused();
-  await page.getByTestId('model-search').fill('synth');
-
-  await expect(page.getByTestId('model-option')).toHaveCount(2);
-  await expect(page.getByTestId('model-option').nth(0)).toHaveAttribute('data-highlighted', 'true');
-
-  await page.keyboard.press('ArrowDown');
-  await expect(page.getByTestId('model-option').nth(1)).toHaveAttribute('data-highlighted', 'true');
-
-  await page.keyboard.press('Enter');
-  await expect(page.getByTestId('model-picker-button')).toHaveText('Synth Code');
-  await expect(page.getByTestId('model-browser')).toHaveCount(0);
-});
