+++ Sources/QuillCodeApp/QuillCodeHitTargetViewModifiers.swift
@@ -1,147 +1,8 @@
 import SwiftUI
 
 extension View {
-    fileprivate func quillCodeInteractiveTarget(_ spec: QuillCodeHitTargetSpec) -> some View {
+    func quillCodeInteractiveTarget(_ spec: QuillCodeHitTargetSpec) -> some View {
         modifier(QuillCodeHitTargetModifier(spec: spec))
-    }
-
-    public func quillCodeTextButtonTarget(
-        minWidth: CGFloat = QuillCodeMetrics.compactTextButtonMinWidth,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> some View {
-        quillCodeInteractiveTarget(.textButton(
-            minWidth: minWidth,
-            minHeight: minHeight,
-            alignment: alignment,
-            radius: radius
-        ))
-    }
-
-    public func quillCodeFormActionTarget(
-        minWidth: CGFloat = QuillCodeMetrics.compactFormActionMinWidth,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center
-    ) -> some View {
-        quillCodeInteractiveTarget(.formAction(
-            minWidth: minWidth,
-            minHeight: minHeight,
-            alignment: alignment
-        ))
-    }
-
-    public func quillCodeTextEntryTarget(
-        minWidth: CGFloat? = nil,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> some View {
-        quillCodeInteractiveTarget(.textEntry(
-            minWidth: minWidth,
-            minHeight: minHeight,
-            alignment: alignment,
-            radius: radius
-        ))
-    }
-
-    public func quillCodeSegmentedControlTarget(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center
-    ) -> some View {
-        quillCodeInteractiveTarget(.segmentedControl(
-            minHeight: minHeight,
-            alignment: alignment
-        ))
-    }
-
-    public func quillCodeAdjustableControlTarget(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center
-    ) -> some View {
-        quillCodeInteractiveTarget(.adjustableControl(
-            minHeight: minHeight,
-            alignment: alignment
-        ))
-    }
-
-    public func quillCodeLinkTarget(
-        minWidth: CGFloat? = QuillCodeMetrics.compactTextButtonMinWidth,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> some View {
-        quillCodeInteractiveTarget(.link(
-            minWidth: minWidth,
-            minHeight: minHeight,
-            alignment: alignment,
-            radius: radius
-        ))
-    }
-
-    public func quillCodeSwitchRowTarget(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .leading
-    ) -> some View {
-        quillCodeInteractiveTarget(.switchRow(
-            minHeight: minHeight,
-            alignment: alignment
-        ))
-    }
-
-    public func quillCodeOwnedGestureTarget(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .leading,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> some View {
-        quillCodeInteractiveTarget(.ownedGesture(
-            minHeight: minHeight,
-            alignment: alignment,
-            radius: radius
-        ))
-        .accessibilityAddTraits(.isButton)
-    }
-
-    public func quillCodeIconButtonTarget(
-        size: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        radius: CGFloat = QuillCodeMetrics.iconControlRadius
-    ) -> some View {
-        quillCodeInteractiveTarget(.icon(size: size, radius: radius))
-    }
-
-    public func quillCodeFullRowButtonTarget(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .leading,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> some View {
-        quillCodeInteractiveTarget(.fullRow(
-            minHeight: minHeight,
-            alignment: alignment,
-            radius: radius
-        ))
-    }
-
-    public func quillCodeCapsuleButtonTarget(
-        minWidth: CGFloat? = nil,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center
-    ) -> some View {
-        quillCodeInteractiveTarget(.capsule(
-            minWidth: minWidth,
-            minHeight: minHeight,
-            alignment: alignment
-        ))
-    }
-
-    public func quillCodeDecorativeIconFrame(
-        size: CGFloat = QuillCodeMetrics.minimumHitTarget
-    ) -> some View {
-        frame(width: size, height: size)
-            .accessibilityHidden(true)
-    }
-
-    public func quillCodePlatformMenuItemTarget(reason: String) -> some View {
-        modifier(QuillCodePlatformMenuItemTargetModifier(reason: reason))
     }
 }
 
@@ -194,7 +55,7 @@ private struct QuillCodeHitTargetModifier: ViewModifier {
     }
 }
 
-private struct QuillCodePlatformMenuItemTargetModifier: ViewModifier {
+struct QuillCodePlatformMenuItemTargetModifier: ViewModifier {
     var reason: String
 
     func body(content: Content) -> some View {
