+++ Sources/QuillCodeApp/WorkspacePullRequestReviewThreadSurface.swift
@@ -91,123 +91,3 @@ public struct WorkspacePullRequestReviewThreadSurface: Codable, Sendable, Hashab
         self.selector = selector
     }
 }
-
-public struct WorkspacePullRequestReviewThreadCommentSurface: Codable, Sendable, Hashable, Identifiable {
-    public var id: String
-    public var databaseID: Int?
-    public var author: String?
-    public var body: String
-
-    public var oneLineBody: String {
-        body
-            .split(whereSeparator: \.isNewline)
-            .map { $0.trimmingCharacters(in: .whitespacesAndNewlines) }
-            .filter { !$0.isEmpty }
-            .joined(separator: " ")
-    }
-
-    public init(
-        id: String,
-        databaseID: Int? = nil,
-        author: String? = nil,
-        body: String
-    ) {
-        self.id = id
-        self.databaseID = databaseID
-        self.author = author
-        self.body = body
-    }
-}
-
-public struct WorkspacePullRequestReviewThreadReplyTargetSurface: Codable, Sendable, Hashable, Identifiable {
-    public var threadID: String
-    public var commentID: Int
-    public var selector: String?
-
-    public var id: String {
-        "reply:\(threadID):\(commentID)"
-    }
-
-    public init(threadID: String, commentID: Int, selector: String? = nil) {
-        self.threadID = threadID
-        self.commentID = commentID
-        self.selector = selector
-    }
-
-    public func request(body: String) -> WorkspacePullRequestReviewThreadReplyRequest? {
-        let trimmedBody = body.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmedBody.isEmpty else { return nil }
-        return WorkspacePullRequestReviewThreadReplyRequest(
-            threadID: threadID,
-            commentID: commentID,
-            body: trimmedBody,
-            selector: selector
-        )
-    }
-}
-
-public struct WorkspacePullRequestReviewThreadReplyRequest: Codable, Sendable, Hashable, Identifiable {
-    public var threadID: String
-    public var commentID: Int
-    public var body: String
-    public var selector: String?
-
-    public var id: String {
-        "reply:\(threadID):\(commentID)"
-    }
-
-    public init(
-        threadID: String,
-        commentID: Int,
-        body: String,
-        selector: String? = nil
-    ) {
-        self.threadID = threadID
-        self.commentID = commentID
-        self.body = body
-        self.selector = selector
-    }
-}
-
-public enum WorkspacePullRequestReviewThreadActionKind: String, Codable, Sendable, Hashable {
-    case resolve
-    case unresolve
-
-    public var title: String {
-        switch self {
-        case .resolve:
-            return "Resolve"
-        case .unresolve:
-            return "Unresolve"
-        }
-    }
-
-    public var systemImage: String {
-        switch self {
-        case .resolve:
-            return "checkmark.circle"
-        case .unresolve:
-            return "arrow.uturn.backward.circle"
-        }
-    }
-}
-
-public struct WorkspacePullRequestReviewThreadActionSurface: Codable, Sendable, Hashable, Identifiable {
-    public var kind: WorkspacePullRequestReviewThreadActionKind
-    public var threadID: String
-    public var selector: String?
-
-    public var id: String {
-        "\(kind.rawValue):\(threadID)"
-    }
-
-    public init(
-        kind: WorkspacePullRequestReviewThreadActionKind,
-        threadID: String,
-        selector: String? = nil
-    ) {
-        self.kind = kind
-        self.threadID = threadID
-        self.selector = selector
-    }
-}
